* add missing const keyword to new_const functions * limit VisualColor and VisuallyEqual traits to testing, move their implementations to separate files and also limit them to testing * remove neg_inf and inf methods from Real trait and add separate traits LowerBounded and UpperBounded * Remove more superfluous trait bounds -- specially Real
