+++ palette/src/macros/equality.rs
@@ -1,9 +1,7 @@
 macro_rules! impl_eq {
-
     (  $self_ty: ident , [$($element: ident),+]) => {
         impl_eq!($self_ty<>, [$($element),+]);
     };
-
     (  $self_ty: ident < $($ty_param: ident),* > , [$($element: ident),+]) => {
 
         impl<$($ty_param,)* T> PartialEq for $self_ty<$($ty_param,)* T>
