+++ palette/src/num.rs
@@ -27,14 +27,6 @@ pub trait Real {
     /// Create a number from an `f64` value, mainly for converting constants.
     #[must_use]
     fn from_f64(n: f64) -> Self;
-
-    /// Returns the negative infinity (−∞).
-    #[must_use]
-    fn neg_inf() -> Self;
-
-    /// Returns the infinity (∞).
-    #[must_use]
-    fn inf() -> Self;
 }
 
 /// Trait for creating a vectorized value from a scalar value.
@@ -61,6 +53,18 @@ pub trait IntoScalarArray<const N: usize>: FromScalar {
     /// Creates an array of scalars from a vectorized value.
     #[must_use]
     fn into_array(self) -> [Self::Scalar; N];
+}
+
+/// Numbers which have lower bounds
+pub trait LowerBounded {
+    /// Returns the smallest finite number this type can represent
+    fn min_value() -> Self;
+}
+
+/// Numbers which have upper bounds
+pub trait UpperBounded {
+    /// Returns the largest finite number this type can represent
+    fn max_value() -> Self;
 }
 
 /// Methods for the value `0`.
@@ -335,6 +339,18 @@ macro_rules! impl_uint {
                 }
             }
 
+            impl LowerBounded for $ty {
+                fn min_value() -> Self{
+                    $ty::MIN
+                }
+            }
+
+            impl UpperBounded for $ty {
+                fn max_value() -> Self{
+                    $ty::MAX
+                }
+            }
+
             impl MinMax for $ty {
                 #[inline]
                 fn min(self, other: Self) -> Self {
@@ -429,13 +445,6 @@ macro_rules! impl_float {
                 fn from_f64(n: f64) -> $ty {
                     n as $ty
                 }
-                #[inline]
-                fn neg_inf()-> $ty{
-                    $ty::NEG_INFINITY
-                }
-                fn inf()-> $ty{
-                    $ty::INFINITY
-                }
             }
 
             impl FromScalar for $ty {
@@ -473,6 +482,18 @@ macro_rules! impl_float {
                 #[inline]
                 fn one() -> Self {
                     1.0
+                }
+            }
+
+            impl LowerBounded for $ty {
+                fn min_value() -> Self{
+                    $ty::MIN
+                }
+            }
+
+            impl UpperBounded for $ty {
+                fn max_value() -> Self{
+                    $ty::MAX
                 }
             }
